+++ logic/leader.lua
@@ -47,16 +47,6 @@ local function KickPlayersAndLeave()
 
             FastReset.Share.EXITINSTANCE.VALUE = 0
 
-            -- send hodor eject 500MS later
-            zo_callLater( function()
-                -- send additional hodor eject
-                if FastReset.savedVariables.sendAdditionalHodorEject then
-                    if HodorReflexes then
-                        HodorReflexes.modules.share.SendCustomData(22, true)
-                    end
-                end
-            end,500)
-
             -- start listener
             EVENT_MANAGER:RegisterForEvent(FastReset.name .. "KickRecieved", EVENT_PLAYER_ACTIVATED, function()
                 EVENT_MANAGER:UnregisterForEvent(FastReset.name .. "KickRecieved", EVENT_PLAYER_ACTIVATED)