+++ modules/share/dataShare.lua
@@ -1,6 +1,11 @@
 FastReset = FastReset or {}
 FastReset.Share = FastReset.Share or {}
 
+local LGB = LibGroupBroadcast
+local _LGBHandler = {}
+local _LGBProtocols = {}
+local MESSAGE_ID_FASTRESET = 35
+
 FastReset.Share = {
     MAXDEATHCOUNT = {
         OFFSET = 0,
@@ -23,142 +28,75 @@ FastReset.Share = {
         VALUE = 0
     },
 
-    MapID = 21, -- Khenarthi's Roost (max data: 40000^2-1)
     DeathCountDelay = 10000,
 }
 
 
-local function rshift(data, by)
-    --FastReset.debug("shift right: " .. data .. " by " .. 2^by .. " - got: " .. zo_floor(data / (2^by)))
-    return zo_floor(data / (2^by))
-end
-local function lshift(data, by)
-    --FastReset.debug("shift left: " .. data .. " by " .. 2^by .. " - got: " .. (data* (2^by)))
-    return (data* (2^by))
-end
+function FastReset.Share:TransmitData(now)
+    local data = {}
+    data.maxDeathCount = FastReset.Share.MAXDEATHCOUNT.VALUE
+    data.deathCount = FastReset.Share.DEATHCOUNT.VALUE
+    data.exitInstance = FastReset.Share.EXITINSTANCE.VALUE
+    data.inNewTrial = FastReset.Share.INNEWTRIAL.VALUE
 
-local function decode(data, offset, size)
-    data = rshift(data, offset)
-    return data % (2^size)
-end
-local function encode(data, offset, size)
-    if size ~= nil then
-        data = data % (2^size)
-    end
-    return lshift(data, offset)
+    _LGBProtocols[MESSAGE_ID_FASTRESET]:Send(data)
 end
 
-local function encodeData(data)
-    if type(data) ~= "table" then return nil end
 
-    local result = 0
-    result = result + encode(data.MAXDEATHCOUNT, FastReset.Share.MAXDEATHCOUNT.OFFSET, FastReset.Share.MAXDEATHCOUNT.SIZE)
-    result = result + encode(data.DEATHCOUNT, FastReset.Share.DEATHCOUNT.OFFSET, FastReset.Share.DEATHCOUNT.SIZE)
-    result = result + encode(data.EXITINSTANCE, FastReset.Share.EXITINSTANCE.OFFSET, FastReset.Share.EXITINSTANCE.SIZE)
-    result = result + encode(data.INNEWTRIAL, FastReset.Share.INNEWTRIAL.OFFSET, FastReset.Share.INNEWTRIAL.SIZE)
+local function messageHandler(unitTag, data)
+    if not FastReset.savedVariables.enabled then return end
+    if not IsUnitGroupLeader(unitTag) then return end
 
-    return result
-end
-local function decodeData(data)
-    local result = {}
-    result.MAXDEATHCOUNT = decode(data, FastReset.Share.MAXDEATHCOUNT.OFFSET, FastReset.Share.MAXDEATHCOUNT.SIZE)
-    result.DEATHCOUNT = decode(data, FastReset.Share.DEATHCOUNT.OFFSET, FastReset.Share.DEATHCOUNT.SIZE)
-    result.EXITINSTANCE = decode(data, FastReset.Share.EXITINSTANCE.OFFSET, FastReset.Share.EXITINSTANCE.SIZE)
-    result.INNEWTRIAL = decode(data, FastReset.Share.INNEWTRIAL.OFFSET, FastReset.Share.INNEWTRIAL.SIZE)
-    return result
-end
+    FastReset.Share.MAXDEATHCOUNT.VALUE = data.maxDeathCount
+    FastReset.Share.DEATHCOUNT.VALUE = data.deathCount
+    FastReset.Share.INNEWTRIAL.VALUE = data.inNewTrial
 
-function FastReset.Share.dataShareHandler(tag, rawData)
-    if not IsUnitGroupLeader(tag) then return end
-
-    local data = decodeData(rawData)
-
-    FastReset.Share.MAXDEATHCOUNT.VALUE = data.MAXDEATHCOUNT
-    FastReset.Share.DEATHCOUNT.VALUE = data.DEATHCOUNT
-    FastReset.Share.INNEWTRIAL.VALUE = data.INNEWTRIAL
-
-    --[[
-    FastReset.debug("MAXDEATHCOUNT: " .. tostring(data.MAXDEATHCOUNT))
-    FastReset.debug("DEATHCOUNT: " .. tostring(data.DEATHCOUNT))
-    FastReset.debug("EXITINSTANCE: " .. tostring(data.EXITINSTANCE))
-    FastReset.debug("INNEWTRIAL: " .. tostring(data.INNEWTRIAL))
-    ]]--
-
-    if data.EXITINSTANCE == 1 then
+    if data.exitInstance == 1 then
         FastReset.Member.KickRecieved()
         --d("recieved exit command")
     end
-    --[[
-    if data.INNEWTRIAL == 1 then
+    if data.inNewTrial == 1 then
         --d("recieved that the leader is in the new trial and we can port to him")
     end
-    ]]--
 end
 
+local function declareLGBProtocols()
+    local CreateNumericField = LGB.CreateNumericField
+    local CreateFlagField = LGB.CreateFlagField
 
+    local protocolOptions = {
+        isRelevantInCombat = true
+    }
 
---function FastReset.share.PrepareData()
---
---end
+    local handler = LGB:RegisterHandler("FastReset")
+    handler:SetDisplayName("FastReset")
+    handler:SetDescription("Allows automatically resetting a trial instance when deaths occur")
 
-local function packData()
-    local rawData = {}
-    rawData.MAXDEATHCOUNT = FastReset.Share.MAXDEATHCOUNT.VALUE
-    rawData.DEATHCOUNT = FastReset.Share.DEATHCOUNT.VALUE
-    rawData.EXITINSTANCE = FastReset.Share.EXITINSTANCE.VALUE
-    rawData.INNEWTRIAL = FastReset.Share.INNEWTRIAL.VALUE
-    return encodeData(rawData)
+    local protocolFastReset = handler:DeclareProtocol(MESSAGE_ID_FASTRESET, "FastReset")
+    protocolFastReset:AddField(CreateNumericField("maxDeathCount", {
+        minValue = 0,
+        maxValue = 127,
+    }))
+    protocolFastReset:AddField(CreateNumericField("deathCount", {
+        minValue = 0,
+        maxValue = 127,
+    }))
+    protocolFastReset:AddField(CreateFlagField("exitInstance", {
+        defaultValue = false,
+    }))
+    protocolFastReset:AddField(CreateFlagField("inNewTrial", {
+        defaultValue = false,
+    }))
+    protocolFastReset:OnData(messageHandler)
+    protocolFastReset:Finalize(protocolOptions)
+
+    _LGBHandler = handler
+    _LGBProtocols[MESSAGE_ID_FASTRESET] = protocolFastReset
 end
-
-function FastReset.Share:TransmitData(now)
-    local data = packData()
-    if not now then
-        self.dataShare:QueueData(data)
-    else
-        self.dataShare:SendData(data)
-    end
-
-    --[[
-    local sentData = decodeData(data)
-    FastReset.debug("sent data: ")
-    FastReset.debug("MAXDEATHCOUNT: " .. sentData.MAXDEATHCOUNT)
-    FastReset.debug("DEATHCOUNT: " .. sentData.DEATHCOUNT)
-    FastReset.debug("EXITINSTANCE: " .. sentData.EXITINSTANCE)
-    FastReset.debug("INNEWTRIAL: " .. sentData.INNEWTRIAL)
-    ]]--
-end
-
---[[
-local function sendTestData()
-    local rawData = {}
-    rawData.MAXDEATHCOUNT = math.random( 1, 63)
-    rawData.DEATHCOUNT = math.random(1, 63)
-    rawData.EXITINSTANCE = math.random(0, 1)
-    rawData.INNEWTRIAL = math.random(0, 1)
-
-    FastReset.debug("sent MAXDEATHCOUNT: " .. tostring(rawData.MAXDEATHCOUNT))
-    FastReset.debug("sent DEATHCOUNT: " .. tostring(rawData.DEATHCOUNT))
-    FastReset.debug("sent EXITINSTANCE: " .. tostring(rawData.EXITINSTANCE))
-    FastReset.debug("sent INNEWTRIAL: " .. tostring(rawData.INNEWTRIAL))
-
-    local encoded = encodeData(rawData)
-    FastReset.debug("encoded data: " .. tostring(encoded))
-
-    FastReset.Share.dataShare:SendData(encoded)
-end
-SLASH_COMMANDS["/frtest"] = function(str)
-    sendTestData()
-end
-
-SLASH_COMMANDS["/frtest"] = function(str)
-    FastReset.Share:TransmitData(true)
-end
-]]--
 
 function FastReset.Share:Register()
-    self.dataShare = LibDataShare:RegisterMap(FastReset.name, self.MapID, self.dataShareHandler)
+    declareLGBProtocols()
 end
 function FastReset.Share:Unregister()
-    LibDataShare:UnregisterMap(FastReset.Share.MapID)
-    self.dataShare = nil
+    d("")
 end